+++ app/src/main/AndroidManifest.xml
@@ -3,7 +3,6 @@
   package="org.oppia.app">
   <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE" />
   <uses-permission android:name="android.permission.INTERNET" />
-  <uses-permission android:name="android.permission.WRITE_EXTERNAL_STORAGE" />
   <!-- TODO(#56): Reenable landscape support. -->
   <uses-sdk android:minSdkVersion="19"
     android:maxSdkVersion="28" />
